bug fix in astor_housing added support for users table and purchases table. also added support for tax tags and access tax tags

+++ astor_real_estate/astor_schemas.py
@@ -57,7 +57,7 @@ class BuildingSchema(Schema):
     otherarea = fields.Int()
     gr_sqft = fields.Int()
     property_tax = fields.Float()
-    numfloors = fields.Int()
+    numfloors = fields.Float()
     unitsres = fields.Int()
     unitstotal = fields.Int()
     yearbuilt = fields.Int()
